istr.h: avoid warning from clang on 'always true' expression
In some cases we use the macro with a char instead of int, which makes
the static analysis true that the condition is pointless.  Add a cast in
the macro to make the argument an int.  Not sure if that makes clang
fully happy though.

Issue: https://github.com/grobian/html2text/issues/35

+++ istr.h
@@ -21,7 +21,7 @@
 
 /* crude temp hack until we properly use wchar, glibc isspace crashes on
  * too large values */
-#define isspace(X) ((X) > 0 && (X) < 256 && isspace(X))
+#define isspace(X) (((int)(X)) > 0 && ((int)(X)) < 256 && isspace((int)(X)))
 
 class istr {
 	public:
